file-struct redux formed, dep redux-actions added
