+++ package.json
@@ -15,6 +15,7 @@
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
     "redux": "^4.0.5",
+    "redux-actions": "^2.6.5",
     "redux-persist": "^6.0.0",
     "web-vitals": "^0.2.4"
   },
