+++ src/redux/actions/index.js
@@ -1,0 +1,5 @@
+import { saveToken, deleteToken } from './token.js';
+import { setUser, resetUser } from './user.js';
+import { setNews } from './news.js';
+
+export { saveToken, deleteToken, setUser, resetUser, setNews }
