+++ src/redux/actions/news.js
@@ -1,0 +1,8 @@
+import { SET_NEWS } from '../../constants/actions-types.js';
+
+export const setNews = (news) => {
+  return {
+    type: SET_NEWS,
+    payload: news
+  }
+}
