+++ src/redux/store/index.js
@@ -1,7 +1,7 @@
 import { createStore } from 'redux';
 import { persistStore, persistReducer } from 'redux-persist';
 import storage from 'redux-persist/lib/storage';
-import rootReducer from '../reducers/index.js';
+import rootReducer from '../../redux/reducers/index.js';
 
 const persistConfig = {
   key: 'root',
